# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/menus": {
            "get": {
                "tags": [
                    "04.菜单接口"
                ],
                "summary": "菜单列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "04.菜单接口"
                ],
                "summary": "新增菜单",
                "parameters": [
                    {
                        "description": "菜单信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/menus/{id}": {
            "put": {
                "tags": [
                    "04.菜单接口"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "04.菜单接口"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}/form": {
            "get": {
                "tags": [
                    "04.菜单接口"
                ],
                "summary": "获取菜单表单数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/notices/{ids}": {
            "delete": {
                "tags": [
                    "08.通知公告"
                ],
                "summary": "删除通知公告",
                "parameters": [
                    {
                        "type": "string",
                        "description": "公告ID列表",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/notices/{id}": {
            "put": {
                "tags": [
                    "08.通知公告"
                ],
                "summary": "修改通知公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/notices/{id}/detail": {
            "get": {
                "tags": [
                    "08.通知公告"
                ],
                "summary": "通知公告详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/notices/{id}/form": {
            "get": {
                "tags": [
                    "08.通知公告"
                ],
                "summary": "通知公告表单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/notices/{id}/publish": {
            "put": {
                "tags": [
                    "08.通知公告"
                ],
                "summary": "发布通知公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/notices/{id}/revoke": {
            "put": {
                "tags": [
                    "08.通知公告"
                ],
                "summary": "撤回通知公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/roles/{id}/menus": {
            "put": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "分配菜单权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单ID列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.KeyValue": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "model.MenuForm": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "alwaysShow": {
                    "type": "integer"
                },
                "component": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keepAlive": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.KeyValue"
                    }
                },
                "parentId": {
                    "type": "integer"
                },
                "perm": {
                    "type": "string"
                },
                "redirect": {
                    "type": "string"
                },
                "routeName": {
                    "type": "string"
                },
                "routePath": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "C",
                        "M",
                        "B"
                    ]
                },
                "visible": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
...
    }
}